Fix description and example
