+++ README.md
@@ -1,2 +1,14 @@
 # distributed-framework
-Simple framework for running distributed algorithms
+A collection of algorithms for _Distributed Systems_ course (winter semester 2020/21) at [Jagiellonian University](https://uj.edu.pl), [Theoretical Computer Science Department](https://tcs.uj.edu.pl).
+
+## Algorithms
+
+### Leader election on synchronized rings
+1. ~~Chang-Roberts algorithm~~
+
+## Running
+
+Run example:
+```bash
+  go run src/example/example_undirected.go 5
+```
